Betty checking

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -6,7 +6,7 @@
 */
 int counting(int a)
 {
-	int counter = 0;
+	long unsigned int counter = 0;
 
 	while (a != 0)
 	{
